🚑 `fs::remove_dir_all` is not idempotent

+++ src/helpers/src/generate.rs
@@ -361,7 +361,9 @@ pub fn trash_outdated_websites() -> Result<State, Error> {
     // NOTE: This whould not be necessary since the directory should be deleted
     //   but there might be edge cases where it's still there and the next
     //   `fs::rename` will fail if it's the case.
-    fs::remove_dir_all(TRASH_DIR.as_path())?;
+    if TRASH_DIR.exists() {
+        fs::remove_dir_all(TRASH_DIR.as_path())?;
+    }
 
     // Remove outdated websites
     fs::rename(DEST_DIR.as_path(), TRASH_DIR.as_path())?;
